Fixes injection point of SubCat
SubCategory def assumed the item designators were added in the correct order, assuming the first was the left most. If this was not the case, and multiple ones got deleted, it would inject the category too far right, or even out of bounds. Added iterative check to set index to leftmost designator.

Also set preventCameraMotion to false, for obvious reasons.